simplified users api routes

+++ api/users.js
@@ -1,56 +1,51 @@
 const prisma = require("../db/prisma");
 const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
 const { userRequired, adminRequired } = require("./utils");
 const userRouter = require("express").Router();
 
-const { JWT_SECRET } = process.env;
 const SALT_ROUNDS = 10;
 
-userRouter.post(
-  "/create",
-  userRequired,
-  adminRequired,
-  async (req, res, next) => {
-    let user;
-    try {
-      const {
-        first_name,
-        last_name,
-        email,
-        preferred_name,
-        gpa,
-        address,
-        phone,
-        password,
-      } = req.body;
-      const hashedPassword = await bcrypt.hash(password, SALT_ROUNDS);
-      user = await prisma.User.create({
-        data: {
-          first_name: first_name,
-          last_name: last_name,
-          email: email,
-          preferred_name: preferred_name,
-          gpa: gpa ? gpa : 4.0,
-          address: address,
-          phone: phone,
-          password: hashedPassword,
-        },
-      });
-    } catch (error) {
-      console.log(error);
-    }
-    if (user) {
-      delete user.password;
-
-      res.send({ user });
-    } else res.send("A user with that email already exists.");
+userRouter.post("/", userRequired, adminRequired, async (req, res, next) => {
+  let user;
+  try {
+    const {
+      first_name,
+      last_name,
+      email,
+      preferred_name,
+      gpa,
+      address,
+      phone,
+      password,
+      is_admin,
+    } = req.body;
+    const hashedPassword = await bcrypt.hash(password, SALT_ROUNDS);
+    user = await prisma.User.create({
+      data: {
+        first_name: first_name,
+        last_name: last_name,
+        email: email,
+        preferred_name: preferred_name,
+        gpa: gpa ? gpa : 4.0,
+        address: address,
+        phone: phone,
+        password: hashedPassword,
+        is_admin: is_admin == "TRUE",
+      },
+    });
+  } catch (error) {
+    console.error(error);
   }
-);
+  if (user) {
+    delete user.password;
+    res.send({ user });
+  } else res.send("A user with that email already exists.");
+});
 
-userRouter.get("/read", userRequired, adminRequired, async (req, res, next) => {
+userRouter.get("/", userRequired, adminRequired, async (req, res, next) => {
   const { email, id } = req.body;
   let user;
+
   try {
     if (id) {
       user = await prisma.User.findUnique({
@@ -64,8 +59,11 @@ userRouter.get("/read", userRequired, adminRequired, async (req, res, next) => {
           email: email,
         },
       });
+    } else {
+      user = req.user;
     }
     if (user) {
+      delete user.password;
       res.send(user);
     } else res.send(`No account found.`);
   } catch (error) {
@@ -73,7 +71,7 @@ userRouter.get("/read", userRequired, adminRequired, async (req, res, next) => {
   }
 });
 
-userRouter.patch("/update", userRequired, async (req, res, next) => {
+userRouter.patch("/", userRequired, async (req, res, next) => {
   try {
     const user = req.user;
     const id = user.id;
@@ -92,6 +90,7 @@ userRouter.patch("/update", userRequired, async (req, res, next) => {
         password: password ? password : user.password,
       },
     });
+    delete updatedUser.password;
     res.send({ updatedUser });
   } catch (error) {
     next(error);
@@ -99,7 +98,7 @@ userRouter.patch("/update", userRequired, async (req, res, next) => {
 });
 
 userRouter.patch(
-  "/update/:id",
+  "/:id",
   userRequired,
   adminRequired,
   async (req, res, next) => {
@@ -140,6 +139,7 @@ userRouter.patch(
             is_admin: is_admin ? is_admin == "TRUE" : user.is_admin,
           },
         });
+        delete updatedUser.password;
         res.send({ updatedUser });
       }
     } catch (error) {
@@ -148,40 +148,40 @@ userRouter.patch(
   }
 );
 
-userRouter.post(
-  "/delete",
-  userRequired,
-  adminRequired,
-  async (req, res, next) => {
-    const { email, id } = req.body;
-    let user;
-    try {
-      if (id) {
-        user = await prisma.User.findUnique({
-          where: {
-            id: +id,
-          },
-        });
-      } else if (email) {
-        user = await prisma.User.findUnique({
-          where: {
-            email: email,
-          },
-        });
-      }
-      if (user) {
+userRouter.delete("/", userRequired, adminRequired, async (req, res, next) => {
+  const { email, id } = req.body;
+  let user;
+  try {
+    if (id) {
+      user = await prisma.User.findUnique({
+        where: {
+          id: +id,
+        },
+      });
+    } else if (email) {
+      user = await prisma.User.findUnique({
+        where: {
+          email: email,
+        },
+      });
+    }
+    if (user) {
+      if (user.id === req.user.id) {
+        res.send("You can't delete yourself, dummy.");
+      } else {
         await prisma.courseUser.deleteMany({ where: { user_id: user.id } });
         const deletedUser = await prisma.User.delete({
           where: {
             id: user.id,
           },
         });
+        delete deletedUser.password;
         res.send(deletedUser);
-      } else res.send(`No account found.`);
-    } catch (error) {
-      next(error);
-    }
+      }
+    } else res.send(`No account found.`);
+  } catch (error) {
+    next(error);
   }
-);
+});
 
 module.exports = userRouter;
